moved data into buffer and texture create info
