+++ src/bx/framework/resources/texture.cpp
@@ -46,7 +46,8 @@ bool Resource<Texture>::Load(const String& filename, Texture& data)
     createInfo.format = TextureFormat::RGBA8_UNORM_SRGB;
     createInfo.size = Extend3D(data.width, data.height, 1);
     createInfo.usageFlags = TextureUsageFlags::TEXTURE_BINDING | TextureUsageFlags::STORAGE_BINDING | TextureUsageFlags::COPY_SRC;
-    data.m_texture = Graphics::CreateTexture(createInfo, static_cast<const void*>(data.pixels.data()));
+    createInfo.data = static_cast<const void*>(data.pixels.data());
+    data.m_texture = Graphics::CreateTexture(createInfo);
 
     return true;
 }
