+++ src/bx/framework/systems/renderer/present_pass.cpp
@@ -1,6 +1,6 @@
 #include "bx/framework/systems/renderer/present_pass.hpp"
 
-#include "bx/engine/containers/lazy_init.hpp"
+#include "bx/framework/systems/renderer/lazy_init.hpp"
 
 const char* PRESENT_SHADER_SRC = R""""(
